Split project and config toml files

+++ external/tgc/tgc.c
@@ -5,7 +5,7 @@ static size_t tgc_hash(void *ptr) {
 }
 
 static size_t tgc_probe(tgc_t* gc, size_t i, size_t h) {
-  long v = i - (h-1);
+  ptrdiff_t v = i - (h-1);
   if (v < 0) { v = gc->nslots + v; }
   return v;
 }
